refactor: remove commenting domain and merge into comment.
* refactor: remove commenting domain and merge into comment.

* refactor: replace @Where annotation with @SQLRestriction for soft delete logic
