--- src/main/java/org/nfactorial/newsfeed/domain/comment/dto/UpdateCommentResponse.java
@@ -1,7 +1,0 @@
-package org.nfactorial.newsfeed.domain.comment.dto;
-
-public record UpdateCommentResponse(
-	long id,
-	String content
-) {
-}
