+++ src/main/java/org/nfactorial/newsfeed/domain/comment/dto/request/UpdateCommentRequest.java
@@ -1,4 +1,4 @@
-package org.nfactorial.newsfeed.domain.comment.dto;
+package org.nfactorial.newsfeed.domain.comment.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.Builder;
