+++ src/main/java/org/nfactorial/newsfeed/domain/comment/dto/response/UpdateCommentResponse.java
@@ -1,0 +1,6 @@
+package org.nfactorial.newsfeed.domain.comment.dto.response;
+
+public record UpdateCommentResponse(
+	long id,
+	String content) {
+}
