+++ src/main/java/org/nfactorial/newsfeed/domain/comment/dto/response/WriteCommentToPostResponse.java
@@ -1,8 +1,8 @@
-package org.nfactorial.newsfeed.domain.commenting.dto.response;
+package org.nfactorial.newsfeed.domain.comment.dto.response;
 
 import java.time.LocalDateTime;
 
-import org.nfactorial.newsfeed.domain.commenting.dto.result.WriteCommentToPostResult;
+import org.nfactorial.newsfeed.domain.comment.dto.result.WriteCommentToPostResult;
 
 import lombok.Builder;
