+++ src/main/java/org/nfactorial/newsfeed/domain/comment/service/CommentService.java
@@ -4,7 +4,6 @@ import java.util.List;
 
 import org.nfactorial.newsfeed.common.code.ErrorCode;
 import org.nfactorial.newsfeed.common.exception.BusinessException;
-import org.nfactorial.newsfeed.domain.comment.dto.WriteCommentCommand;
 import org.nfactorial.newsfeed.domain.comment.entity.Comment;
 import org.nfactorial.newsfeed.domain.comment.repository.CommentRepository;
 import org.nfactorial.newsfeed.domain.post.entity.Post;
@@ -25,14 +24,6 @@ public class CommentService implements CommentServiceApi {
 	}
 
 	@Transactional
-	@Override
-	public Comment writeComment(WriteCommentCommand command) {
-		Comment comment = Comment.write(command.post(), command.profile(), command.content());
-		Comment savedComment = commentRepository.save(comment);
-		return savedComment;
-	}
-
-	@Transactional
 	public void deleteById(long commentId, long profileId) {
 		Comment comment = getOwnedComment(commentId, profileId);
 		commentRepository.delete(comment);
@@ -43,13 +34,6 @@ public class CommentService implements CommentServiceApi {
 		Comment comment = getOwnedComment(commentId, profileId);
 		comment.updateContent(content);
 		return comment.getContent();
-	}
-
-	@Override
-	@Transactional(readOnly = true)
-	public List<Comment> commentListByPost(Post post) {
-		List<Comment> comments = commentRepository.findAllByPost(post);
-		return comments;
 	}
 
 	private Comment getOwnedComment(long commentId, long profileId) {
