+++ src/main/java/org/nfactorial/newsfeed/domain/comment/service/PostCommentingService.java
@@ -1,13 +1,12 @@
-package org.nfactorial.newsfeed.domain.commenting.service;
+package org.nfactorial.newsfeed.domain.comment.service;
 
 import java.util.List;
 
-import org.nfactorial.newsfeed.domain.comment.dto.WriteCommentCommand;
+import org.nfactorial.newsfeed.domain.comment.dto.command.WriteCommentToPostCommand;
 import org.nfactorial.newsfeed.domain.comment.entity.Comment;
-import org.nfactorial.newsfeed.domain.comment.service.CommentServiceApi;
-import org.nfactorial.newsfeed.domain.commenting.dto.command.WriteCommentToPostCommand;
-import org.nfactorial.newsfeed.domain.commenting.dto.result.CommentListByPostResult;
-import org.nfactorial.newsfeed.domain.commenting.dto.result.WriteCommentToPostResult;
+import org.nfactorial.newsfeed.domain.comment.repository.CommentRepository;
+import org.nfactorial.newsfeed.domain.comment.dto.result.CommentListByPostResult;
+import org.nfactorial.newsfeed.domain.comment.dto.result.WriteCommentToPostResult;
 import org.nfactorial.newsfeed.domain.post.entity.Post;
 import org.nfactorial.newsfeed.domain.post.service.PostServiceApi;
 import org.nfactorial.newsfeed.domain.profile.entity.Profile;
@@ -19,16 +18,17 @@ import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
-public class PostCommentService {
+public class PostCommentingService {
     private final PostServiceApi postService;
     private final ProfileServiceApi profileService;
-    private final CommentServiceApi commentService;
+    private final CommentRepository commentRepository;
 
     @Transactional
     public WriteCommentToPostResult writeCommentToPost(WriteCommentToPostCommand command) {
         Post post = postService.getPostById(command.postId());
         Profile profile = profileService.getProfileEntityById(command.profileId());
-        Comment savedComment = commentService.writeComment(WriteCommentCommand.of(post, profile, command.content()));
+        Comment comment = Comment.write(post, profile, command.content());
+        Comment savedComment = commentRepository.save(comment);
         return WriteCommentToPostResult.builder()
             .id(savedComment.getId())
             .createdAt(savedComment.getCreatedAt())
@@ -39,7 +39,7 @@ public class PostCommentService {
     @Transactional(readOnly = true)
     public CommentListByPostResult commentListByPost(long postId) {
         Post post = postService.getPostById(postId);
-        List<Comment> comments = commentService.commentListByPost(post);
+        List<Comment> comments = commentRepository.findAllByPost(post);
         return CommentListByPostResult.of(comments);
     }
 }
