+++ src/test/java/org/nfactorial/newsfeed/domain/comment/CommentE2ETest.java
@@ -9,8 +9,8 @@ import org.nfactorial.newsfeed.domain.auth.dto.LoginRequest;
 import org.nfactorial.newsfeed.domain.auth.dto.LoginResponse;
 import org.nfactorial.newsfeed.domain.auth.dto.SignUpRequest;
 import org.nfactorial.newsfeed.domain.comment.repository.CommentRepository;
-import org.nfactorial.newsfeed.domain.commenting.dto.request.WriteCommentToPostRequest;
-import org.nfactorial.newsfeed.domain.commenting.dto.response.WriteCommentToPostResponse;
+import org.nfactorial.newsfeed.domain.comment.dto.request.WriteCommentToPostRequest;
+import org.nfactorial.newsfeed.domain.comment.dto.response.WriteCommentToPostResponse;
 import org.nfactorial.newsfeed.domain.post.dto.request.PostCreateRequest;
 import org.nfactorial.newsfeed.domain.post.dto.response.PostCreateResponse;
 import org.nfactorial.newsfeed.domain.post.repository.PostRepository;
