+++ src/test/java/org/nfactorial/newsfeed/domain/comment/UpdateCommentE2ETest.java
@@ -9,8 +9,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.nfactorial.newsfeed.common.dto.GlobalApiResponse;
-import org.nfactorial.newsfeed.domain.comment.dto.UpdateCommentRequest;
-import org.nfactorial.newsfeed.domain.comment.dto.UpdateCommentResponse;
+import org.nfactorial.newsfeed.domain.comment.dto.request.UpdateCommentRequest;
+import org.nfactorial.newsfeed.domain.comment.dto.response.UpdateCommentResponse;
 import org.nfactorial.newsfeed.domain.comment.entity.Comment;
 import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.http.HttpEntity;
